Moved merge funtionality from the data quality app to a new data manager app.
